allocate webport at node-create, not at runtime
This simplifies a lot of code, as the baseurl is known before the Node
is even created. The create-node CLI arguments were changed.

Also, webapi tokens are cleared earlier, in Node.init_webport(), since
the previous sequence wasn't avoiding the race condition mentioned in
the comments (which wasn't a real race anyways, since the reactor never
gets control at a time that would let the uncleared state be
observable).