use EventSource for relay client
changed format of regular non-EventSource GET to be the same (e.g. data:
foo\n\n)

modified EventSource to handle non-"text/event-source" responses,
assuming they're in the same format as an event-source would deliver.

invitation.py: disabled re-sending of messages that we think the server
has lost. Now that the events are delivered one at a time, this logic is
flawed.