command_invite(): add offer_mailbox=, accept mailbox offer upon receipt
If offer_mailbox=True, the sender will allocate a transport and include
the necessary details in the invitation payload. When the other agent
receives the payload, it will store the details and begin advertising
the new transport. (this is the "auto-accept" mode, which should be made
more deliberate in the future)