schema: "STRING" is not legal SQL, use "VARCHAR" instead
SQLite behaves unfortunately when it doesn't recognize the column type:
it appears to inspect the value being stored to decide upon a type. When
you store a string like "0123", you get back an int/long like 123.
"STRING" is not recognized, so when my hex-encoded tokens happened to
not include any A-F characters, they got mangled on the way back out.
This was pretty rare until I made one of the tokens (RT) shorter, at
which point it started happening maybe 10% of the time and I finally
noticed it.

"VARCHAR" usually requires a max-length (e.g. VARCHAR(100)), but sqlite
appears to be happy without it.