setup.py: make it somewhat pip-installable
Adds packages=, because it turns out that versioneer.py doesn't play
well with projects that lack a packages= setting (the
PACKAGE/_version.py file that it wants to rewrite doesn't get installed
without one).

Adds an entry_points= that points to a new function in runner.py .

Adds MANIFEST.in entries to include all the non-py files (SQL in
db-schema/ and HTML/JS/CSS in media/).

This isn't completely pip-installable yet, as our dependencies aren't
expressed in a way that pip understands. I'm still trying to decide how
I want to handle that (how to make them secure).