create addressbook entry at start of invitation process
(instead of at its conclusion).

This cleans up the invitation database entry, as we no longer need a
place to stash private info about the invitation (e.g. the long-term
keys we'll be using). This also helps the frontend render pending- and
completed- invitations in the same way: they're all contacts, but some
are more ready for use than others.