provide control of the 'accept-mailbox-offer' flag
This adds a checkbox to the control-panel details view, where you can
toggle acceptance of mailbox offers back and forth. At present, this
only affects things if you set it before the invitation is complete, but
in the future it will let you stop using a previously-accepted mailbox,
or accept an offer after the fact.