remove safe_develop, use plain pip install
I still like the idea of safe_develop, and it should be easier to implement
now (since peep's hash-checking was merged into pip at some point), but at
the moment the pinned dependencies are causing problems. In particular, the
current versions of cffi and pynacl build correctly on most systems, but the
older versions pinned in requirements*.txt don't.